correct_il/envs/f1/requirements.txt
gym==0.19.0
matplotlib==3.7.2
numpy==1.22.0
pygame==2.5.0
PyYAML==6.0.1
tabulate==0.9.0
torch==2.0.1
git+https://github.com/f1tenth/f1tenth_gym.git@cd56335eda43ff4e401331c461877227474a3ed4#egg=f110_gym